backend/requirements.txt
flask
flask-sqlalchemy
python-dotenv
werkzeug
flask-jwt-extended
google-genai
razorpay